Promote ReferenceRepositoryType to official RepositoryType in graphql.  Update specs and sencondary references
