--- app/graphql/types/reference_repository_edge_type.rb
@@ -1,6 +1,0 @@
-
-# frozen_string_literal: true
-
-class ReferenceRepositoryEdgeType < GraphQL::Types::Relay::BaseEdge
-  node_type(ReferenceRepositoryType)
-end
