+++ app/graphql/types/repository_connection_with_total_type.rb
@@ -5,12 +5,10 @@ class RepositoryConnectionWithTotalType < BaseConnection
   field_class GraphQL::Cache::Field
 
   field :total_count, Integer, null: true, cache: true
-  field :years, [FacetType], null: true, cache: true
-  field :members, [FacetType], null: true, cache: true
   field :software, [FacetType], null: true, cache: true
   field :certificates, [FacetType], null: true, cache: true
-  field :clientTypes, [FacetType], null: true, cache: true
-  field :repositoryTypes, [FacetType], null: true, cache: true
+  field :repository_types, [FacetType], null: true, cache: true
+  field :years, [FacetType], null: true, cache: true
 
   def total_count
     object.total_count
@@ -20,23 +18,15 @@ class RepositoryConnectionWithTotalType < BaseConnection
     facet_by_year(object.aggregations.years.buckets)
   end
 
-  def members
-    facet_by_combined_key(object.aggregations.providers.buckets)
-  end
-
   def software
     facet_by_software(object.aggregations.software.buckets)
   end
 
-  def certificates
-    facet_by_key(object.aggregations.certificates.buckets)
-  end
-
-  def client_types
-    facet_by_key(object.aggregations.client_types.buckets)
-  end
-
   def repository_types
     facet_by_key(object.aggregations.repository_types.buckets)
+  end
+
+  def certificates
+    facet_by_key(object.aggregations.certificates.buckets)
   end
 end
