+++ app/graphql/types/repository_edge_type.rb
@@ -1,3 +1,4 @@
+
 # frozen_string_literal: true
 
 class RepositoryEdgeType < GraphQL::Types::Relay::BaseEdge
